# openapi.yaml
...
openapi: 3.1.0
info:
  description: Invoices and payments for each business. Amounts are whole US cents.
  title: Invoice and Payment Service
  version: 1.0.0
servers:
- url: /
security:
- apiKey: []
paths:
...
  /api/v1/events:
    get:
      operationId: listEvents
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageEvent"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Webhooks and events
...
  /api/v1/invoices/{id}/pay:
    post:
      description: Starts a payment and finishes it in the background. Check the payment
        status or wait for a webhook. Sending the same request again returns the same
        answer.
      operationId: payInvoice
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - description: A unique value for this payment. Send the same value if you retry.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
          pattern: "^[!-~]{1,128}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentAccepted"
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not find this item.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "The invoice is paid, a payment is in progress, or this request\
            \ was used before."
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: This request is too large.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      summary: Pay an invoice
      tags:
      - Payments
  /api/v1/invoices/{id}/payment-attempts:
    get:
      operationId: listPaymentAttempts
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagePaymentAttempt"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not find this item.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Payments
  /api/v1/payment-attempts/{id}:
    get:
      operationId: getPaymentAttempt
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentAttempt"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not find this item.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Payments
...
  /api/v1/webhook-endpoints:
    get:
      operationId: listWebhookEndpoints
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageWebhookEndpoint"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Webhooks and events
    post:
      operationId: registerWebhookEndpoint
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewWebhookEndpoint"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisteredWebhookEndpoint"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: This request is too large.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Some details are missing or not valid.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Webhooks and events
  /api/v1/webhook-endpoints/{id}:
    get:
      operationId: getWebhookEndpoint
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not read this request.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Use a valid API key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: We could not find this item.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Something went wrong. Please try again.
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: The service is busy. Please try again.
      tags:
      - Webhooks and events
components:
  schemas:
...
    ErrorDetails:
      properties:
        code:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/InvalidField"
        message:
          type: string
        request_id:
          type: string
    ErrorResponse:
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetails"
    Event:
      properties:
        created_at:
          type: string
          format: date-time
        event_type:
          type: string
          enum:
          - invoice.created
          - invoice.paid
          - invoice.payment_failed
        id:
          type: string
          format: uuid
        payload:
          type: object
    InvalidField:
      properties:
        field:
          type: string
        message:
          type: string
...
    NewWebhookEndpoint:
      properties:
        url:
          type: string
          maxLength: 2048
          minLength: 1
      required:
      - url
...
    PageEvent:
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        next_cursor:
          type: string
...
    PagePaymentAttempt:
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PaymentAttempt"
        next_cursor:
          type: string
...
    PageWebhookEndpoint:
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEndpoint"
        next_cursor:
          type: string
    PaymentAccepted:
      properties:
        invoice_id:
          type: string
          format: uuid
        payment_attempt_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - pending
          - unknown
          - succeeded
          - failed
        status_url:
          type: string
    PaymentAttempt:
      properties:
        amount_cents:
          type: integer
          format: int64
        amount_display:
          type: string
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        currency:
          type: string
        failure_code:
          type: string
        id:
          type: string
          format: uuid
        invoice_id:
          type: string
          format: uuid
        last_error_code:
          type: string
        psp_reference:
          type: string
        review_required:
          type: boolean
        status:
          type: string
          enum:
          - pending
          - unknown
          - succeeded
          - failed
    PaymentRequest:
      properties:
        card_token:
          type: string
          minLength: 1
          pattern: tok_(success|insufficient_funds|card_declined|timeout|network_error)
      required:
      - card_token
    RegisteredWebhookEndpoint:
      properties:
        active:
          type: boolean
        id:
          type: string
          format: uuid
        signing_secret:
          type: string
        url:
          type: string
...
    WebhookEndpoint:
      properties:
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        url:
          type: string
  securitySchemes:
    apiKey:
      description: Use a valid API key.
      scheme: bearer
      type: http